+++ @yundeving/story-book-demo-ui/ui/components/Button/Button.type.ts
@@ -1,10 +1,10 @@
 import { type VariantType, type SizeType } from "../common.type";
 
-export type ButtonState = "default" | "loading" | "success" | "error";
+export type ButtonStatus = "default" | "loading" | "success" | "error";
 
 export interface ButtonProps extends React.ComponentProps<"button"> {
   label?: string;
   variant?: VariantType;
   size?: SizeType;
-  state?: ButtonState;
+  status?: ButtonStatus;
 }