+++ @yundeving/story-book-demo-ui/ui/components/Card/Card.tsx
@@ -33,7 +33,10 @@ const CardTitle = ({ className, ...props }: React.ComponentProps<"div">) => {
   return (
     <div
       data-slot="card-title"
-      className={cn("leading-none font-semibold", className)}
+      className={cn(
+        "text-default-800 dark:text-default-200 leading-none font-semibold",
+        className
+      )}
       {...props}
     />
   );