+++ @yundeving/story-book-demo-ui/ui/components/StatusButton/StatusButton.type.ts
@@ -1,5 +1,5 @@
 import { type ButtonProps } from "../Button/Button.type";
 
 export interface StatusButtonProps extends ButtonProps {
-  state: "loading" | "success" | "error";
+  status: "loading" | "success" | "error";
 }